# Optimize flaky_workflows_jobs CH query