### Changes

  1. Pre-filtering tables before joins:
    - Added filtering conditions to the workflow_run subquery before joining with workflow_job
    - Reduced data volume by filtering early rather than after joins
  2. Simplified data structure:
    - Separated the original CTE into two steps (dedups_wo_push and latest_attempts)
    - The push table is joined only after initial filtering, reducing join complexity
  3. Optimized conclusion handling:
    - Replaced string comparisons with numeric codes (0 for success, 1 for failure, 2 for other)
    - This improves comparison performance in window functions
  4. Improved partitioning:
    - Changed partition keys from concatenated string "fullname" to separate workflow_name and job_name columns
  5. Added test cases:
    - Added specific test parameters to params.json for validation testing



### Perf

```
+------+----------+-----------+-------------+---------------+------------+------------+-------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change |  Avg Mem   |  Base Mem  |  Mem Change | % Mem Change |
+------+----------+-----------+-------------+---------------+------------+------------+-------------+--------------+
|  0   |   3151   |    None   |     None    |      None     | 3629068029 |    None    |     None    |     None     |
|  1   |   3832   |   11943   |    -8111    |      -68      | 3626686195 | 7653666021 | -4026979826 |     -53      |
+------+----------+-----------+-------------+---------------+------------+------------+-------------+--------------+
```
note: base version often exceeds memory limits



### Testing
```
./clickhouse_query_perf.py --query flaky_workflows_jobs --results --times 1  --verbose --base HEAD

./clickhouse_query_perf.py --query flaky_workflows_jobs --results --times 1  --verbose --base HEAD
Using unstaged changes for --head
Comparing results for query: flaky_workflows_jobs
Num tests: 2
Head: unstaged
 Base: HEAD

=== Test 0 New Results (first 1 rows) ===
Row 0: {
  "workflow_name": "pull",
  "workflow_id": "14072313906",
  "job_name": "linux-focal-cuda12.6-py3.10-gcc11 / test (default, 5, 5, lf.linux.4xlarge.nvidia.gpu)",
  "job_id": "39410432865",
  "flaky": 1,
  "run_attempt": "1",
  "next_workflow_id": "14072313906",
  "next_job_id": "39410432865",
  "annotation": ""
}
Total rows: 1
==================================================
Results for test 0 match
Base query for test 1 failed: HTTPDriver for https://hyt81izu0c.us-east-1.aws.clickhouse.cloud:8443 received ClickHouse error code 241

=== Test 1 New Results (first 1 rows) ===
Row 0: {
  "workflow_name": "pull",
  "workflow_id": "14094503541",
  "job_name": "linux-jammy-py3.10-clang15-asan / test (default, 1, 6, linux.4xlarge)",
  "job_id": "39479895763",
  "flaky": 1,
  "run_attempt": "1",
  "next_workflow_id": "14094503541",
  "next_job_id": "39479895763",
  "annotation": ""
}
Total rows: 1
==================================================
Cannot compare results for test 1 - base query failed
New results can be found in the _logs/query_results folder

```
(it's a bit hard to test, as the base query often exceeds memory limits, but both test results match)